Remove shape abstraction (for now) and updated types
