+++ src/shapes/Rectangle.ts
@@ -1,13 +1,65 @@
-import Shape from './Shape';
+import CanvasManager from '../CanvasManager';
 
-class Rectangle extends Shape {
-  constructor(props, parent) {
-    super(props, parent);
+type RectangleProps = {
+  x: number;
+  y: number;
+  z: number;
+  width: number;
+  height: number;
+  color: string;
+}
+
+interface Rectangle {
+  props: RectangleProps;
+  parent: CanvasManager;
+  draw(): void;
+}
+
+class Rectangle {
+  constructor(props: RectangleProps, parent: CanvasManager) {
+    this.props = props;
+    this.parent = parent;
 
     this.draw();
   }
 
-  get color() {
+  get x():number {
+    return this.props.x;
+  }
+
+  set x(x) {
+    this.props.x = x;
+    this.parent.redraw();
+  }
+
+  get y():number {
+    return this.props.y;
+  }
+
+  set y(y) {
+    this.props.y = y;
+    this.parent.redraw();
+  }
+
+  get width():number {
+    return this.props.width;
+  }
+
+  set width(width) {
+    this.props.width = width;
+    this.parent.redraw();
+  }
+
+  get height():number {
+    return this.props.height;
+  }
+
+  set height(height) {
+    this.props.height = height;
+    this.parent.redraw();
+  }
+
+  get color():string {
     return this.props.color;
   }
 
@@ -22,16 +74,4 @@ class Rectangle extends Shape {
   }
 }
 
-interface Rectangle extends Shape {
-  props: {
-    x: number;
-    y: number;
-    z: number;
-    width: number;
-    height: number;
-    color: string;
-  }
-  draw(): void;
-}
-
-export default Rectangle;
+export { Rectangle as default, RectangleProps };
